requirements.txt
annotated-types==0.5.0
ansiwrap==0.8.4
anyio==3.7.1
appnope==0.1.3
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
cjio==0.8.2
click==8.1.7
cms==0.0.2
colorama==0.4.6
cryptography==41.0.4
cssselect2==0.7.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-admin-interface==0.26.0
django-check-seo==0.6.4
django-ckeditor==6.7.0
django-classy-tags==4.1.0
django-cms==3.11.4
django-colorfield==0.9.0
django-filer==3.0.6
django-formtools==2.4.1
django-js-asset==2.1.0
django-meta==2.3.0
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-simple-menu==2.1.2.post1
django-treebeard==4.7
djangocms-admin-style==3.2.5
djangocms-page-meta==1.4.0
docker==6.1.3
dsdev-utils==1.3.0
easy-thumbnails==2.8.5
entrypoints==0.4
escapism==1.0.1
executing==1.2.0
fastjsonschema==2.18.0
halo==0.0.31
html5lib==1.1
humanfriendly==10.0
idna==3.4
ipykernel==6.10.0
ipython==8.14.0
ipython-genutils==0.2.0
iso8601==2.0.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client==6.1.5
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter-repo2docker==2023.6.0
jupyter-server==1.15.6
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
log-symbols==0.0.14
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
Menus==0.2.0
mistune==3.0.1
nbclassic==1.0.0
nbclient==0.7.0
nbconvert==7.7.4
nbformat==5.9.2
nbstripout==0.6.1
nest-asyncio==1.5.7
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.25.2
oscrypto==1.3.0
packaging==23.1
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.16.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
PyYAML==6.0.1
pyzmq==24.0.1
qrcode==7.4.2
referencing==0.30.2
reportlab==3.6.13
requests==2.31.0
rpds-py==0.9.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semver==3.0.1
Send2Trash==1.8.2
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spinners==0.0.24
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
textwrap3==0.9.2
timeago==1.0.16
tinycss2==1.2.1
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
treebeard==0.1.1
typing_extensions==4.7.1
tzlocal==5.0.1
Unidecode==1.3.6
uritools==4.0.2
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.2
wrapt==1.15.0
xhtml2pdf==0.2.11
